# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Booking:
    properties:
      alternatives:
        items:
          $ref: '#/definitions/domain.BookingAlternative'
        type: array
      comment:
        type: string
      completed_at:
        type: string
      confirmed_at:
        type: string
      created_at:
        type: string
      date:
        type: string
      duration:
        type: integer
      guests_count:
        type: integer
      id:
        type: string
      rejected_at:
        type: string
      restaurant_id:
        type: string
      status:
        $ref: '#/definitions/domain.BookingStatus'
      time:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  domain.BookingAlternative:
    properties:
      accepted_at:
        type: string
      booking_id:
        type: string
      created_at:
        type: string
      date:
        type: string
      id:
        type: string
      message:
        type: string
      rejected_at:
        type: string
      time:
        type: string
    type: object
  domain.BookingStatus:
    enum:
    - pending
    - confirmed
    - rejected
    - cancelled
    - completed
    type: string
    x-enum-varnames:
    - BookingStatusPending
    - BookingStatusConfirmed
    - BookingStatusRejected
    - BookingStatusCancelled
    - BookingStatusCompleted
...
  domain.Notification:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_read:
        type: boolean
      message:
        type: string
      recipient_id:
        type: string
      recipient_type:
        $ref: '#/definitions/domain.RecipientType'
      related_id:
        type: string
      title:
        type: string
      type:
        $ref: '#/definitions/domain.NotificationType'
    type: object
  domain.NotificationType:
    enum:
    - new_booking
    - booking_confirmed
    - booking_rejected
    - booking_cancelled
    - alternative_offer
    - alternative_accepted
    - alternative_rejected
    type: string
    x-enum-varnames:
    - NotificationTypeNewBooking
    - NotificationTypeBookingConfirmed
    - NotificationTypeBookingRejected
    - NotificationTypeBookingCancelled
    - NotificationTypeAlternativeOffer
    - NotificationTypeAlternativeAccepted
    - NotificationTypeAlternativeRejected
  domain.RecipientType:
    enum:
    - user
    - restaurant
    type: string
    x-enum-varnames:
    - RecipientTypeUser
    - RecipientTypeRestaurant
...
  domain.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
...
  handlers.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    required:
    - email
    - name
    - phone
    type: object
...
  handlers.UpdateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    required:
    - email
    - name
    - phone
    type: object
host: localhost:8080
info:
  contact:
    email: support@restaurant-booking.com
    name: API Support
  description: API для системы бронирования ресторанов
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Restaurant Booking API
  version: "1.0"
paths:
...
  /users:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Invalid data
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Email already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create user
      tags:
      - users
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Get user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update an existing user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Invalid data
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Email already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update user
      tags:
      - users
  /users/{id}/bookings:
    get:
      consumes:
      - application/json
      description: Get all bookings of a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Booking'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user bookings
      tags:
      - users
      - bookings
  /users/{id}/notifications:
    get:
      consumes:
      - application/json
      description: Get all notifications of a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Notification'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user notifications
      tags:
      - users
      - notifications
schemes:
- http
- https
swagger: "2.0"